+++ library/src/main/java/io/fastpix/media3/info/FastPixPlayerLibraryInfo.kt
@@ -1,0 +1,6 @@
+package io.fastpix.media3.info
+
+object FastPixPlayerLibraryInfo {
+    const val PLAYER_NAME = "fastpix-android-player"
+    const val PLAYER_VERSION = "1.0.7"
+}